refactor(backend-worker): move files
