+++ backend-worker/src/main/kotlin/org/rucca/snake/worker/SubmitController.kt
@@ -1,11 +1,10 @@
-package org.rucca.snake.worker.controller
+package org.rucca.snake.worker
 
 import org.rucca.cheese.auth.AuthenticationService
 import org.rucca.cheese.auth.annotation.Guard
 import org.rucca.snake.worker.api.SubmitApi
 import org.rucca.snake.worker.model.SubmitPost200ResponseDTO
 import org.rucca.snake.worker.model.SubmitPost200ResponseDataDTO
-import org.rucca.snake.worker.service.SubmitService
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.RestController
 import org.springframework.web.multipart.MultipartFile
